test: move new test with flaky since it not work on current pipeline
